refactor: Unify file naming
