+++ packages/aws-lambda-otel-extension/opt/otel-extension/external/proto/metric-service.proto
